requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
blis==1.3.0
cachetools==6.2.1
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
et_xmlfile==2.0.0
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
gliner==0.2.22
hf-xet==1.1.10
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.45.1
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.8.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.1
openai-whisper==20250625
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.3.1
preshed==3.0.10
protobuf==6.32.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydantic==2.12.1
pydantic_core==2.41.3
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==5.1.1
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sklearn==0.0
smart_open==7.3.1
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.14.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.35.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wordcloud==1.9.4
wrapt==1.17.3
